pyproject.toml
[tool.poetry]
name = "icrm"
version = "0.1.0"
description = "Pet project to learn viewflow."
authors = ["Denys Zaluzhnyi <denys.zaluzhnyi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
django-viewflow = "^2.1.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
